refactor: rename gk-detailed to complex
