+++ src/page-components/complex/DetailedInfo.tsx
